[fix] For long standing image shifting bug
Fixed shifting window start point when redisplaying message.  Image,
above the redisplayed message, has had tend to shift to the left, so
when you press for example on video message and video file starts
downloading, Image (if any) above this message started to move to the
left.  I thought it was some nasty bug in Emacs redisplay, but this
was caused by tweaking around window-start point

This is fixed by this commit