Use telega-tracking as the new face for tracking notifications
When adding a buffer for notifications with the tracking package, one
can specify a list of faces to use for it in the modeline.  This is
useful not only for appearance, but also because the face can be used
to prioritize the notification, altering the order in which it is
shown (via tracking-faces-priorities).