[enh] Do no query user on `telega-open-file'
When file associated with Telegram message is opened for the long time
when you try to reopen it again Emacs will query you "Reread file from
disk?" because modtime of the buffer and modtime of newly downloaded
file differs.  This commit workarounds this by setting buffer's
modtime to file's one in case buffer is not modified.